Add ability to generate a random point on a circle
- Added functionality to generate a random point on a circle. This
allows us to do some extra constraint filtering. For example, if
constraints cannot be filled because one point must be on two lines, but
that's the only constraint on that point, we can automatically place it.
This is extended to circles where if there is one circle that we need to
place a point on, we can now generate that.